requirements.txt
argparse
colorama
termcolor
pathlib
pycryptodome
py7zr